requirements.txt
scikit-learn~=1.7.1
pdfplumber~=0.11.7
python-docx~=1.2.0
fastapi~=0.116.1
pydantic-settings~=2.10.1
faiss-cpu~=1.12.0
numpy~=2.3.2
pandas~=2.3.2
protobuf~=5.29.5